Refactor into cargo workspace
