+++ Cargo.toml
@@ -1,47 +1,9 @@
-[package]
-name = "libaris-rs"
-version = "0.1.0"
-authors = ["Avi Weinstock <aweinstock314@gmail.com>"]
+[workspace]
 
-[dependencies]
-base64 = "0.12.0"
-#either = "1.5.0"
-failure = "0.1.6"
-frunk = "0.2.2"
-lazy_static = "1.4.0"
-nom = "4.1.1"
-petgraph = "0.4.13"
-sha2 = "0.8.1"
-varisat = "0.2.1"
-xml-rs = "0.8.0"
-itertools = "0.9.0"
-strum = "0.18.0"
-strum_macros = "0.18.0"
-
-jni = { version = "0.10.2", optional = true }
-yew = { version = "0.16.2", optional = true }
-gloo = { version = "0.2.0", optional = true }
-wasm-bindgen = { version = "^0.2", optional = true }
-wee_alloc = { version = "0.4", optional = true }
-js-sys = { version = "0.3", optional = true }
-web-sys = { version = "0.3", optional = true }
-
-[features]
-java = ["jni"]
-js = ["js-sys", "web-sys", "web-sys/HtmlAnchorElement", "yew/web_sys", "wasm-bindgen", "wee_alloc/size_classes", "gloo"]
-workaround_rls_vscode_bug_586 = []
-
-[build-dependencies]
-cbindgen = "*"
-version_check = "0.9"
-
-[lib]
-name = "libaris"
-crate_type = ["cdylib", "rlib"]
-
-[[bin]]
-name = "auto_grader"
-path = "src/bin/auto_grader.rs"
-
-[profile.release]
-debug = true # enable debugging symbols in release mode (primarily for line numbers for RUST_BACKTRACE=1)
+members = [
+    "aris",
+    "web-app",
+    "auto-grader",
+    "bindings/c",
+    "bindings/java",
+]
