+++ aris/src/combinatorics.rs
