+++ aris/src/equivalences.rs
