+++ aris/src/parser.rs
@@ -108,7 +108,7 @@ fn test_parser() {
 /// Convert ASCII characters and macros to logic symbols.
 ///
 /// ```rust
-/// # use libaris::parser::prettify_expr;
+/// # use aris::parser::prettify_expr;
 /// assert_eq!(prettify_expr("_|_ -> (^|^ .bicon ~P)"), "⊥ → (⊤ ↔ ¬P)");
 /// ```
 pub fn prettify_expr(s: &str) -> String {
