+++ aris/src/solver_integration/solver.rs
