+++ example-proofs/propositional_logic_arguments_for_proofs_ii_problem_10.bram
