+++ example-proofs/resolution_example.bram
