+++ web-app/src/lib.rs
@@ -1,0 +1,14 @@
+#![recursion_limit="512"]
+
+mod ui;
+
+use wasm_bindgen::prelude::*;
+
+#[wasm_bindgen]
+pub fn run_app() -> Result<(), JsValue> {
+    #[global_allocator]
+    static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
+
+    yew::start_app::<ui::App>();
+    Ok(())
+}
